bugfix syndata
- converted Faraday OP intensities to volts (oops)
- convert (using compose) OP intensities to scientific notation to match Isolynx output
- de-correct ion counter data for deadtime, use this to generate Poisson noise, then re-correct for deadtime.

The thinking behind the last point: "If the real ion beam size on an ion counter with 90% gain and a 30 ns deadtime is 1 Mcps, then it will only produce a response for 900 kcps of that signal.  Of those counts, the IC will be ‘dead’ for some, and it only actually registers 876 kcps.  The software then multiplies this by 1.027 = 1/(1+dt*876000) to get 900 kcps, which is what gets reported.  So, the stats on the beam should be Poisson distributed with a variance of 876,000.  I think.  And a mean of 876,000 * 1.027 = 900,000."